requirements/dev.txt
-r base.txt
django-debug-toolbar==1.0.1
# Required by django-debug-toolbar
sqlparse==0.1.11

coverage==3.7.1
pylint==1.1.0
# Required by pylint
logilab-common==0.61.0
astroid==1.0.1

PyYAML==3.10
Fabric==1.8.2
# Required by Fabric
paramiko==1.13.0
pycrypto==2.6.1
ecdsa==0.11

factory-boy==2.3.1
mock==1.0.1

django-leaflet==0.13.5
